requirements.txt
asgiref==3.8.1
Django==4.2.11
djangorestframework==3.15.1
pillow==10.3.0
pip==20.2.3
setuptools==49.2.1
sqlparse==0.4.4
typing-extensions==4.11.0
